requirements.txt
2captcha-python==1.2.5
amazoncaptcha==0.5.11
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2
django-environ==0.11.2
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.1
drf-yasg==1.21.7
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
lxml==5.2.2
outcome==1.3.0.post0
packaging==24.0
Pillow==9.5.0
pluggy==1.5.0
psycopg2==2.9.9
pycparser==2.22
PySocks==1.7.1
pytest==8.2.1
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.30.0
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
tomli==2.0.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
